Splitting into modules has been performed modified:   index.html new file:   js/data.js modified:   js/main.js new file:   js/util.js
