+++ js/util.js
@@ -1,0 +1,30 @@
+let urlNumber = 0;
+let id = 0;
+
+const getRandomPositiveInteger = (a, b) => {
+  // Если переданы отрицительные числа, возвращаем NaN
+  if (a < 0 || b < 0) {
+    return NaN;
+  }
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+const getUrl = () => {
+  urlNumber += 1;
+  return urlNumber;
+};
+
+const getId = () => {
+  id += 1;
+  return id;
+};
+
+export {getRandomArrayElement};
+export {getUrl};
+export {getId};
+export {getRandomPositiveInteger};
